chore (move Directory) : BowlCount, frame -> pins directory
