+++ src/main/java/bowling/domain/pin/BowlCount.java
@@ -1,4 +1,4 @@
-package bowling.domain.frame;
+package bowling.domain.pin;
 
 import bowling.exception.BowlCountOutOfRangeException;
